refactor: Set interfaces iauth and ihttp
